# src/main/resources/swagger/specs/user.yaml
openapi: 3.0.3
info:
  title: EMS User API
  version: 1.0.0
  description: Endpoints for user authentication, public event discovery, event registrations, teams, payments, and Razorpay webhooks.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/events/{eventId}:
    get:
      tags: [events]
      summary: Get public event by id
      operationId: getEventById
      parameters:
        - $ref: '#/components/parameters/EventId'
      responses:
        '200':
          description: Event details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        default:
          $ref: '#/components/responses/Error'
...
  /api/events/{eventId}/teams:
    get:
      tags: [teams]
      summary: List teams for an event
      operationId: listEventTeams
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
      responses:
        '200':
          description: Teams for the event.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamResponse'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [teams]
      summary: Create a team for an event
      operationId: createEventTeam
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeamRequest'
      responses:
        '200':
          description: Team created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        default:
          $ref: '#/components/responses/Error'
  /api/teams/me:
    get:
      tags: [teams]
      summary: List current user's teams
      operationId: listMyTeams
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Teams containing the current user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamResponse'
        default:
          $ref: '#/components/responses/Error'
  /api/team-requests/me:
    get:
      tags: [teams]
      summary: List current user's pending team requests
      operationId: listMyTeamRequests
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Pending requests and invitations for the current user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamMembershipResponse'
        default:
          $ref: '#/components/responses/Error'
  /api/events/{eventId}/teams/{teamId}:
    get:
      tags: [teams]
      summary: Get team by id
      operationId: getEventTeam
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
        - $ref: '#/components/parameters/TeamId'
      responses:
        '200':
          description: Team details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        default:
          $ref: '#/components/responses/Error'
    put:
      tags: [teams]
      summary: Update a team
      operationId: updateEventTeam
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
        - $ref: '#/components/parameters/TeamId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTeamRequest'
      responses:
        '200':
          description: Team updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [teams]
      summary: Delete a team
      operationId: deleteEventTeam
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
        - $ref: '#/components/parameters/TeamId'
      responses:
        '204':
          description: Team deleted.
        default:
          $ref: '#/components/responses/Error'
  /api/events/{eventId}/teams/{teamId}/members:
    get:
      tags: [teams]
      summary: List team members
      operationId: listEventTeamMembers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
        - $ref: '#/components/parameters/TeamId'
      responses:
        '200':
          description: Team members.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        default:
          $ref: '#/components/responses/Error'
  /api/events/{eventId}/teams/{teamId}/members/{userId}:
    delete:
      tags: [teams]
      summary: Remove a user from a team
      operationId: removeEventTeamMember
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
        - $ref: '#/components/parameters/TeamId'
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: User removed from team.
        default:
          $ref: '#/components/responses/Error'
  /api/events/{eventId}/teams/{teamId}/owner/{userId}:
    put:
      tags: [teams]
      summary: Transfer team ownership
      operationId: transferEventTeamOwnership
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
        - $ref: '#/components/parameters/TeamId'
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Ownership transferred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        default:
          $ref: '#/components/responses/Error'
  /api/events/{eventId}/teams/{teamId}/requests:
    get:
      tags: [teams]
      summary: List pending join requests for a team
      operationId: listEventTeamRequests
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
        - $ref: '#/components/parameters/TeamId'
      responses:
        '200':
          description: Pending team join requests.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamMembershipResponse'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [teams]
      summary: Request to join a team
      operationId: requestToJoinEventTeam
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
        - $ref: '#/components/parameters/TeamId'
      responses:
        '200':
          description: Join request created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamMembershipResponse'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [teams]
      summary: Delete current user's join request
      operationId: deleteMyEventTeamRequest
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
        - $ref: '#/components/parameters/TeamId'
      responses:
        '204':
          description: Join request deleted.
        default:
          $ref: '#/components/responses/Error'
  /api/events/{eventId}/teams/{teamId}/requests/{userId}:
    put:
      tags: [teams]
      summary: Accept or reject a team join request
      operationId: respondToEventTeamRequest
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
        - $ref: '#/components/parameters/TeamId'
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RespondToRequestRequest'
      responses:
        '200':
          description: Join request updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamMembershipResponse'
        default:
          $ref: '#/components/responses/Error'
  /api/events/{eventId}/teams/{teamId}/invitations:
    post:
      tags: [teams]
      summary: Invite a user to a team
      operationId: inviteUserToEventTeam
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
        - $ref: '#/components/parameters/TeamId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteUserRequest'
      responses:
        '200':
          description: Invitation created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamMembershipResponse'
        default:
          $ref: '#/components/responses/Error'
    put:
      tags: [teams]
      summary: Accept or reject current user's team invitation
      operationId: respondToEventTeamInvitation
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
        - $ref: '#/components/parameters/TeamId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RespondToRequestRequest'
      responses:
        '200':
          description: Invitation updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamMembershipResponse'
        default:
          $ref: '#/components/responses/Error'
  /api/events/{eventId}/teams/{teamId}/invitations/{userId}:
    delete:
      tags: [teams]
      summary: Delete a team invitation
      operationId: deleteEventTeamInvitation
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
        - $ref: '#/components/parameters/TeamId'
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: Invitation deleted.
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    EventId:
      in: path
      name: eventId
      required: true
      schema:
        type: integer
        format: int64
    TeamId:
      in: path
      name: teamId
      required: true
      schema:
        type: integer
        format: int64
    UserId:
      in: path
      name: userId
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
  schemas:
    Role:
      type: string
      enum: [USER, MODERATOR, ADMIN]
    EventType:
      type: string
      enum: [SOLO, TEAM]
    EventVisibility:
      type: string
      enum: [PRIVATE, PUBLIC]
    EventStatus:
      type: string
      enum: [DRAFT, PUBLISHED, COMPLETED, CANCELLED]
    RequestStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED, CANCELLED]
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        name:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        emailVerified:
          type: boolean
        phoneVerified:
          type: boolean
    EventResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        createdBy:
          type: integer
          format: int64
        type:
          $ref: '#/components/schemas/EventType'
        visibility:
          $ref: '#/components/schemas/EventVisibility'
        status:
          $ref: '#/components/schemas/EventStatus'
        maxParticipants:
          type: integer
          minimum: 1
        minTeamSize:
          type: integer
          minimum: 1
          nullable: true
        maxTeamSize:
          type: integer
          minimum: 1
          nullable: true
        registrationFee:
          type: number
          format: double
          minimum: 0
        registrationDeadline:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
    TeamResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        ownerId:
          type: integer
          format: int64
        eventId:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TeamMembershipResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        teamId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        type:
          $ref: '#/components/schemas/RequestStatus'
...
    CreateTeamRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
    UpdateTeamRequest:
      type: object
      properties:
        name:
          type: string
    InviteUserRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    RespondToRequestRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [APPROVED, REJECTED]
...
    ApiErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        code:
          type: string
        message:
          type: string
        path:
          type: string
          nullable: true
        details:
          type: array
          items:
            type: string